🐛 fix : 알림 표시 버그 수정 및 store 생성성
