+++ src/routes/Notifications/NotiNav.tsx
@@ -1,0 +1,70 @@
+import { useEffect, useState } from "react";
+import Buttons from "../../components/common/SquareButton";
+import { useNotificationsStore } from "../../store/notificationsStore";
+
+export default function NotiNav() {
+  const { isSeen, isNseenList, getNotificationList, update } =
+    useNotificationsStore();
+  const [init, setInit] = useState<boolean>(true);
+  const [active, setActive] = useState<boolean>(true);
+
+  // 전체 / 읽지 않음 구분 상태
+  const isActive = () => {
+    setActive(() => !active);
+  };
+
+  // 전체 리스트 보기
+  const isYseenList = () => {
+    getNotificationList!();
+  };
+
+  // 모두 읽음 클릭시 읽음처리 실시간 반영을 위한 처리
+  useEffect(() => {
+    // 초기실행 방지지
+    if (init) {
+      setInit(false);
+      return;
+    }
+    getNotificationList!();
+  }, [update]);
+
+  return (
+    <div className="flex w-4/6 justify-between">
+      <div className="flex gap-2 mb-4">
+        <Buttons
+          variant="primary"
+          size="xs"
+          textSize="sm"
+          className="font-normal hover:bg-[#96ccd6]"
+          style={{
+            backgroundColor: active === true ? "#7EACB5" : "#e5e7eb",
+          }}
+          onClick={() => {
+            isActive();
+            isYseenList();
+          }}
+        >
+          전체
+        </Buttons>
+        <Buttons
+          variant="primary"
+          size="xs"
+          textSize="sm"
+          className="font-normal hover:bg-[#96ccd6]"
+          style={{
+            backgroundColor: active === false ? "#7EACB5" : "#e5e7eb",
+          }}
+          onClick={() => {
+            isActive();
+            isNseenList!();
+          }}
+        >
+          읽지 않음
+        </Buttons>
+      </div>
+      <span className="cursor-pointer" onClick={() => isSeen!()}>
+        모두 읽음
+      </span>
+    </div>
+  );
+}
